# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CityPulse urban intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /citypulse/reports:
    get:
      operationId: listCityReports
      tags: [citypulse]
      security:
        - bearerAuth: []
      summary: List accessible reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createCityReport
      tags: [citypulse]
      security:
        - bearerAuth: []
      summary: Create a citizen report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /citypulse/reports/analyze:
    post:
      operationId: analyzeCityReport
      tags: [citypulse]
      security:
        - bearerAuth: []
      summary: Analyze a report with AI decision support
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportAnalysisInput"
      responses:
        "200":
          description: AI-assisted analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportAnalysis"
  /citypulse/reports/{reportId}:
    get:
      operationId: getCityReport
      tags: [citypulse]
      security:
        - bearerAuth: []
      summary: Get one accessible report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /citypulse/reports/{reportId}/status:
    patch:
      operationId: updateCityReportStatus
      tags: [citypulse]
      security:
        - bearerAuth: []
      summary: Apply an administrator report action
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportActionInput"
      responses:
        "200":
          description: Updated report and recorded action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportActionResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      required: [id, user_id, category, title, description, severity, status, created_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: ["string", "null"], format: uuid }
        category: { type: string }
        title: { type: string }
        description: { type: string }
        image_url: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        location_name: { type: ["string", "null"] }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        status:
          type: string
          enum: [PENDING_VERIFICATION, VERIFIED, REJECTED, ASSIGNED, IN_PROGRESS, RESOLVED]
        ai_confidence: { type: ["number", "null"] }
        ai_category: { type: ["string", "null"] }
        ai_summary: { type: ["string", "null"] }
        suspicious: { type: boolean }
        duplicate_of: { type: ["string", "null"], format: uuid }
        evidence_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: ["string", "null"], format: date-time }
    ReportCreateInput:
      type: object
      required: [category, title, description, severity]
      properties:
        category: { type: string, enum: [Traffic, Flooding, "Road Damage", Garbage, Streetlight, Pollution, Infrastructure] }
        title: { type: string, minLength: 4, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 2000 }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        image_url: { type: ["string", "null"] }
        latitude: { type: ["number", "null"], minimum: -90, maximum: 90 }
        longitude: { type: ["number", "null"], minimum: -180, maximum: 180 }
        location_name: { type: ["string", "null"], maxLength: 200 }
        ai_confidence: { type: ["number", "null"], minimum: 0, maximum: 1 }
        ai_category: { type: ["string", "null"] }
        ai_summary: { type: ["string", "null"], maxLength: 500 }
        suspicious: { type: boolean }
        duplicate_of: { type: ["string", "null"], format: uuid }
    ReportAnalysisInput:
      type: object
      required: [category, title, description, severity]
      properties:
        category: { type: string }
        title: { type: string, minLength: 4, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 2000 }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        location_name: { type: ["string", "null"] }
    ReportAnalysis:
      type: object
      required: [category, severity, confidence, summary, suspicious, possible_duplicate, verification_priority]
      properties:
        category: { type: string }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        summary: { type: string }
        suspicious: { type: boolean }
        possible_duplicate: { type: boolean }
        duplicate_of: { type: ["string", "null"], format: uuid }
        verification_priority: { type: string, enum: [LOW, MEDIUM, HIGH] }
        source: { type: string, enum: [gemini, fallback] }
    ReportActionInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [VERIFIED, REJECTED, ASSIGNED, IN_PROGRESS, RESOLVED]
        notes: { type: string, maxLength: 1000 }
    ReportActionResult:
      type: object
      required: [report, action]
      properties:
        report: { $ref: "#/components/schemas/Report" }
        action:
          type: object
          required: [id, admin_id, report_id, action, created_at]
          properties:
            id: { type: string, format: uuid }
            admin_id: { type: string, format: uuid }
            report_id: { type: string, format: uuid }
            action: { type: string }
            notes: { type: ["string", "null"] }
            created_at: { type: string, format: date-time }
...