# lib/api-zod/src/generated/types/reportActionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Report } from './report';
import type { ReportActionResultAction } from './reportActionResultAction';

export interface ReportActionResult {
  report: Report;
  action: ReportActionResultAction;
}