# lib/api-zod/src/generated/types/reportAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportAnalysisSeverity } from './reportAnalysisSeverity';
import type { ReportAnalysisSource } from './reportAnalysisSource';
import type { ReportAnalysisVerificationPriority } from './reportAnalysisVerificationPriority';

export interface ReportAnalysis {
  category: string;
  severity: ReportAnalysisSeverity;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  summary: string;
  suspicious: boolean;
  possible_duplicate: boolean;
  /** @nullable */
  duplicate_of?: string | null;
  verification_priority: ReportAnalysisVerificationPriority;
  source?: ReportAnalysisSource;
}