# lib/api-zod/src/generated/types/reportCreateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportCreateInputCategory } from './reportCreateInputCategory';
import type { ReportCreateInputSeverity } from './reportCreateInputSeverity';

export interface ReportCreateInput {
  category: ReportCreateInputCategory;
  /**
     * @minLength 4
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  severity: ReportCreateInputSeverity;
  /** @nullable */
  image_url?: string | null;
  /**
     * @minimum -90
     * @maximum 90
     * @nullable
     */
  latitude?: number | null;
  /**
     * @minimum -180
     * @maximum 180
     * @nullable
     */
  longitude?: number | null;
  /**
     * @maxLength 200
     * @nullable
     */
  location_name?: string | null;
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  ai_confidence?: number | null;
  /** @nullable */
  ai_category?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  ai_summary?: string | null;
  suspicious?: boolean;
  /** @nullable */
  duplicate_of?: string | null;
}